converted the callbacks beforeDelete and afterDelete for the blog, webpage model into a behavior called ManyToManyCountable.
The callbacks were initially used for updating the
LinkList counter for Links whenever a Blog or Webpage gets deleted
and it has a related Link in the system.

it is built mostly from the Many2ManyCounterCache behavior
currently in use for Product model.
It is for the triage of Product, ProductsInGroup and ProductGroup